requirements_clean.txt
Flask==2.0.1
Werkzeug==2.0.1
azure-cosmos==4.3.0
boto3==1.26.137
python-dotenv==0.19.0
pytest==7.3.1
gunicorn==20.1.0
pdf2image==1.16.3
PyMuPDF==1.20.0
pytesseract==0.3.10
Pillow==9.5.0
pyodbc==4.0.39
requests==2.31.0
